main: add `-format` flag for timestamp format.
I felt like the full format was getting pretty noisy, since I rarely need to
query this information programmatically anyway. Being able to skim it and just
see the timestamps sounds more useful. The date information is available anyway
in the file name. The only information that is lost is the time zone, which is
not really useful anyway.