main: optionally include header with current date and timezone.
The header is included by default; use `-include_header=false` to turn it off.
Also change default time format to just be the 24h clock, without date and
timezone.

The reason for this change is that when I'm `cat`ing together a bunch of snip
files, it becomes pretty difficult to overview. The timestamps are very detailed
and noisy, and it's difficult to glance where a day starts and ends. Including a
header at the top of the file, and making the timestamp format relative to that
header, makes it much easier to overview.